requirements.txt
Django==5.2.8
psycopg2-binary==2.9.9

